Add softmax cross entropy utility
